% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron.R
\docType{data}
\name{chron}
\alias{chron}
\title{Chronological presence / absence data}
\format{
A data frame with 248 rows and 17 variables:
\describe{
\item{OBJECTID}{Unique ID generated by ArcGIS. Not used}
\item{Survey_Uni}{Unique Identifier for each unit generated by the MTAP}
\item{Prehist}{Prehistoric period}
\item{IA}{Iron Age period}
\item{Arch}{Archaic period}
\item{Clas}{Classical period}
\item{Hell}{Hellenistic period}
\item{Cl_He}{Classical to Hellenistic period. Includes all Clas and Hell
and less precisely dated material}
\item{He_ER}{Hellenistic to Early Roman period. Includes all Hell and ER
and less precisely dated material}
\item{ER}{Early Roman period}
\item{MR}{Mid-Roman period}
\item{LR}{Late Roman period}
\item{LR_EB}{Late Roman to Early Byzantine period. Contains all LR and EB
and less precisely dated material}
\item{Roma}{Roman. Conains all material from any Roman period}
\item{EB}{Early Byzantine period}
\item{Byz}{Byzantine period}
\item{Otto}{Ottoman period}
}
}
\source{
Molyvoti, Thrace, Archaeology Project
}
\usage{
chron
}
\description{
A dataset containing presence / absence data for 248 of the units surveyed
in the 2015 season of the Molyvoti, Thrace Archaeology Project (MTAP).
}
\keyword{datasets}
