% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idw_nnear.R
\name{IDW_nnear}
\alias{IDW_nnear}
\title{Inverse Distance Weighting with N Nearest Neighbors}
\usage{
IDW_nnear(tracts, tracts_ID, n)
}
\arguments{
\item{tracts}{A spatial object of class \code{sf}}

\item{tracts_ID}{Character, the name of the ID column in \code{tracts}}

\item{n}{Integer, the number of nearest neighbors}
}
\value{
returns a named list of named lists. Each list is named for a tract
in \code{tracts} and contains \code{n} elements. These elements, named for
the \code{n} nearest neighbors, are the inverse of the distance to the
tract plus one.
}
\description{
Calculate the inverse distance to the n nearest neighbors of any tract
}
\details{
This function uses \code{st_distance} to calculate the distances between all
tracts in an sf object, selects the n nearest neighbors (excluding the tract
itself), calculates the inverse of the distance + 1 (to avoid dividing by 0),
and converts the results into a named list of named lists for use in
\code{\link{NP}}.
}
\examples{
IDW_nnear(tracts = tracts, tracts_ID = "UnitID", n = 8)
}
\seealso{
See \code{\link[sf]{st_distance}} function from package \code{sf}
}
\author{
Eli Weaverdyck \email{eweaverdyck@gmail.com}
}
