% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{NP_colors}
\alias{NP_colors}
\title{Default color scheme for NPCA results}
\usage{
NP_colors(NP_Results)
}
\arguments{
\item{NP_Results}{A character vector. Must be one of the columns generated by \code{\link{NP}} ending in "_Res"}
}
\value{
Returns a character vector of length \code{NP_Results} containing hexadecimal color codes
}
\description{
Applies a default color scheme for displaying the results of NPCA generated by \code{\link{NP}}
}
\examples{
NP_colors(
NP(chron = chron,
  chron_ID = "Survey_Uni",
  periods = c("Clas", "Hell", "Cl_He", "ER", "MR", "LR", "LR_EB"),
  tracts = tracts,
  tracts_ID = "UnitID",
  swl = IDW_nnear(tracts = tracts, tracts_ID = "UnitID", n = 8),
  perms = 100,
  cut = 0.05)[["Clas_Res"]])
}
\seealso{
See \code{\link{NP}}
}
\author{
Eli Weaverdyck \email{eweaverdyck@gmail.com}
}
