% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts.R
\docType{data}
\name{tracts}
\alias{tracts}
\title{Survey units.}
\format{
An \code{sf} object with 248 rows and one variable:
\describe{
\item{UnitID}{Unique Identifier for each unit generated by the MTAP}
}
}
\source{
Molyvoti, Thrace, Archaeology Project
}
\usage{
tracts
}
\description{
A data set containing the geometry for 248 units surveyed in the 2015
season of the Molyvoti, Thrace Archaeology Project (MTAP).
}
\keyword{datasets}
